Fixes, mostly building, for Fedora Core 11

+++ dyninstAPI_RT/src/RTthread-index.c
@@ -236,7 +236,7 @@ unsigned DYNINST_alloc_index(dyntid_t tid)
    result = tc_lock_lock(&DYNINST_index_lock);
    if (result == DYNINST_DEAD_LOCK) {
        rtdebug_printf("%s[%d]:  DEADLOCK HERE tid %lu \n", __FILE__, __LINE__, dyn_pthread_self());
-       fprintf(stderr," %s[%d]:  DEADLOCK HERE tid %lu \n", __FILE__, __LINE__, dyn_pthread_self());
+       fprintf(stderr," %s[%d]:  DEADLOCK HERE tid %lu \n", __FILE__, __LINE__, (unsigned long) dyn_pthread_self());
       /* ERROR_HANDLING_BAD */
       return DYNINST_max_num_threads;
       }
@@ -302,7 +302,7 @@ int DYNINST_free_index(dyntid_t tid)
    result = tc_lock_lock(&DYNINST_index_lock);
    if (result == DYNINST_DEAD_LOCK) {
        rtdebug_printf("%s[%d]:  DEADLOCK HERE tid %lu \n", __FILE__, __LINE__, dyn_pthread_self());
-       fprintf(stderr," %s[%d]:  DEADLOCK HERE tid %lu \n", __FILE__, __LINE__, dyn_pthread_self());
+       fprintf(stderr," %s[%d]:  DEADLOCK HERE tid %lu \n", __FILE__, __LINE__, (unsigned long) dyn_pthread_self());
       return -1;
    }
